+++ voxel/voxel.go
@@ -7,14 +7,14 @@ import (
 type VoxelGrid struct {
 	Parent     *VoxelGrid // higher res version
 	Child      *VoxelGrid // lower res version
-	NumVoxelsX int
-	NumVoxelsY int
-	NumVoxelsZ int
+	NumVoxelsX int32
+	NumVoxelsY int32
+	NumVoxelsZ int32
 	VoxelSize  float32
 	Voxels     []uint8
 }
 
-func NewVoxelGrid(nx, ny, nz int, sz float32) *VoxelGrid {
+func NewVoxelGrid(nx, ny, nz int32, sz float32) *VoxelGrid {
 	return &VoxelGrid{
 		NumVoxelsX: nx,
 		NumVoxelsY: ny,
@@ -24,20 +24,20 @@ func NewVoxelGrid(nx, ny, nz int, sz float32) *VoxelGrid {
 	}
 }
 
-func voxelBitMask(x, y, z int) uint8 {
+func voxelBitMask(x, y, z int32) uint8 {
 	bit := uint8(x + (z * 2) + (y * 2 * 2))
 	mask := uint8(1) << bit
 	return mask
 }
 
-func (grid *VoxelGrid) VoxelIndex(x, y, z int) int {
+func (grid *VoxelGrid) VoxelIndex(x, y, z int32) int32 {
 	vx, vy, vz := x/2, y/2, z/2
 	return vx + vz*grid.NumVoxelsX/2 + vy*grid.NumVoxelsX/2*grid.NumVoxelsZ/2
 }
 
-func (grid *VoxelGrid) GetVoxel(x, y, z int) bool {
+func (grid *VoxelGrid) GetVoxel(x, y, z int32) bool {
 	i := grid.VoxelIndex(x, y, z)
-	if i < 0 || i >= len(grid.Voxels) {
+	if i < 0 || i >= int32(len(grid.Voxels)) {
 		fmt.Printf("GetVoxel: Invalid XYZ - %d, %d, %d\n", x, y, z)
 		return false
 	}
@@ -52,9 +52,9 @@ func (grid *VoxelGrid) GetVoxel(x, y, z int) bool {
 	return voxel&mask != 0
 }
 
-func (grid *VoxelGrid) SetVoxel(x, y, z int, set bool) {
+func (grid *VoxelGrid) SetVoxel(x, y, z int32, set bool) {
 	i := grid.VoxelIndex(x, y, z)
-	if i < 0 || i >= len(grid.Voxels) {
+	if i < 0 || i >= int32(len(grid.Voxels)) {
 		fmt.Printf("SetVoxel: Invalid XYZ - %d, %d, %d\n", x, y, z)
 		return
 	}
@@ -81,9 +81,9 @@ func (grid *VoxelGrid) Compress() *VoxelGrid {
 	newGrid.Parent = grid
 	grid.Child = newGrid
 
-	for y := 0; y < newGrid.NumVoxelsY; y++ {
-		for z := 0; z < newGrid.NumVoxelsZ; z++ {
-			for x := 0; x < newGrid.NumVoxelsX; x++ {
+	for y := int32(0); y < newGrid.NumVoxelsY; y++ {
+		for z := int32(0); z < newGrid.NumVoxelsZ; z++ {
+			for x := int32(0); x < newGrid.NumVoxelsX; x++ {
 
 				px, py, pz := x*2, y*2, z*2
